requirements.txt
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
boto3==1.39.3
botocore==1.39.3
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
django==5.2.4
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
exceptiongroup==1.3.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.7
openai==1.93.1
pip==25.1.1
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pymysql==1.1.1
pyopenssl==25.1.0
pyparsing==3.2.3
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyu2f==0.1.5
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
zstandard==0.23.0
